refactor(sdk): Move persistence implementations to platform-specific directories
- Move `BrowserPersistence` and `ServerPersistence` to `_browser` and `_nodejs`
directories respectively.
- Rename both classes to `Persistence` with named exports, enabling consistent imports via the `@/` path alias that resolves to the appropriate platform directory at build time.
