+++ packages/sdk/src/PersistenceManager.ts
@@ -3,8 +3,8 @@ import { inject, Lifecycle, scoped } from 'tsyringe'
 import { Identity, IdentityInjectionToken } from './identity/Identity'
 import { DestroySignal } from './DestroySignal'
 import { LoggerFactory } from './utils/LoggerFactory'
-import type { Persistence } from './utils/persistence/types'
-import ServerPersistence from './utils/persistence/ServerPersistence'
+import type { Persistence as PersistenceInterface } from './utils/persistence/types'
+import { Persistence } from '@/Persistence'
 
 export const NAMESPACES = {
     ENCRYPTION_KEYS: 'EncryptionKeys',
@@ -14,7 +14,7 @@ export const NAMESPACES = {
 @scoped(Lifecycle.ContainerScoped)
 export class PersistenceManager {
 
-    private persistence?: ServerPersistence
+    private persistence?: Persistence
     private readonly identity: Identity
     private readonly loggerFactory: LoggerFactory
 
@@ -34,7 +34,7 @@ export class PersistenceManager {
     }
 
     private async ensureInitialized() {
-        this.persistence ??= await ServerPersistence.createInstance({
+        this.persistence ??= await Persistence.createInstance({
             loggerFactory: this.loggerFactory,
             ownerId: await this.identity.getUserId(),
             namespaces: Object.values(NAMESPACES),
@@ -42,7 +42,7 @@ export class PersistenceManager {
         })
     }
 
-    async getPersistence(namespace: string): Promise<Persistence> {
+    async getPersistence(namespace: string): Promise<PersistenceInterface> {
         await this.ensureInitialized()
         return {
             get: (key: string): Promise<string | undefined> => {
