+++ packages/sdk/src/_browser/Persistence.ts
@@ -1,5 +1,5 @@
 import { openDB, IDBPDatabase } from 'idb'
-import { PersistenceContext, PersistenceContextOptions } from './PersistenceContext.js'
+import { PersistenceContext, PersistenceContextOptions } from '../utils/persistence/PersistenceContext'
 
 /**
  * This file is a ES module (.mts) instead of CommonJS. It was converted to ESM to resolve
@@ -16,17 +16,17 @@ import { PersistenceContext, PersistenceContextOptions } from './PersistenceCont
  * See https://github.com/streamr-dev/network/pull/2848
  */
 
-export default class BrowserPersistence implements PersistenceContext {
+export class Persistence implements PersistenceContext {
     
     private readonly db: IDBPDatabase
 
-    static async createInstance(opts: PersistenceContextOptions): Promise<BrowserPersistence> {
+    static async createInstance(opts: PersistenceContextOptions): Promise<Persistence> {
         const db = await openDB(`streamr-sdk::${opts.ownerId}`, 1, {
             upgrade(db) {
                 opts.namespaces.forEach((namespace) => db.createObjectStore(namespace))
             }
         })
-        return new BrowserPersistence(db)
+        return new Persistence(db)
     }
 
     private constructor(db: IDBPDatabase) {
